converted to Java class
